# openapi.yaml
openapi: 3.0.0
info:
  title: User Authentication and Account API
  version: 1.0.0
  description: API for user sign-up, sign-in, email verification, password management, and session handling.
servers:
  - url: /api # Assuming your API routes are under /api
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: A human-readable error message.
        details: # Optional, for more specific error info
          type: string
          nullable: true
...
    CategoryBrief:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the category.
        name:
          type: string
          description: Name of the category.
        icon:
          type: string
          nullable: true
          description: Icon for the category (e.g., SVG content or a class name).
      required:
        - id
        - name
    SubcategoryBrief:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the subcategory.
        name:
          type: string
          description: Name of the subcategory.
        categoryId:
          type: string
          description: ID of the parent category.
      required:
        - id
        - name
        - categoryId
    AddonBrief:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        image:
          type: string
          format: url
          nullable: true
        categoryId:
          type: string
      required:
        - id
        - name
        - price
        - categoryId
    PackageBrief:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        image:
          type: string
          format: url
          nullable: true
        categoryId:
          type: string
        subcategoryId:
          type: string
        popular:
          type: boolean
          nullable: true
        bgColor:
          type: string
          nullable: true
      required:
        - id
        - name
        - price
        - categoryId
        - subcategoryId
    Category: # Updated to be CategoryFullResponse
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
          nullable: true
        subcategories:
          type: array
          items:
            $ref: '#/components/schemas/SubcategoryBrief'
        addons:
          type: array
          items:
            $ref: '#/components/schemas/AddonBrief'
        packages:
          type: array
          items:
            $ref: '#/components/schemas/PackageBrief'
      required:
        - id
        - name
...
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          description: New name for the category.
        icon:
          type: string
          nullable: true
          description: New icon for the category.
...
    Addon: # Updated to be AddonFullResponse
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        image:
          type: string
          format: url
          nullable: true
        categoryId:
          type: string
        category:
          $ref: '#/components/schemas/CategoryBrief'
      required:
        - id
        - name
        - price
        - categoryId
    CreateAddonRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        image:
          type: string
          format: url
          nullable: true
        categoryId:
          type: string
      required:
        - name
        - price
        - categoryId
    UpdateAddonRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        image:
          type: string
          format: url
          nullable: true
        categoryId:
          type: string
...
  securitySchemes:
    cookieAuth: # For NextAuth.js session cookies
      type: apiKey
      in: cookie
      name: next-auth.session-token # Or __Secure-next-auth.session-token for HTTPS
security:
  - cookieAuth: [] # Indicates that most authenticated routes use cookieAuth
paths:
...
  /product/categories/{categoryId}:
    parameters:
      - name: categoryId
        in: path
        required: true
        description: ID of the category.
        schema:
          type: string
    get:
      summary: Get a specific category by ID
      tags: [Products]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successfully retrieved the category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category' # Returns full category object
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Category not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a category
      tags: [Products]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: Category updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category' # Returns the updated full category object
        '400':
          description: Invalid input or category has related items preventing update.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Category not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict (e.g., category with this name already exists).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a category
      tags: [Products]
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Category deleted successfully.
        '400':
          description: Category cannot be deleted (e.g., has associated subcategories, packages, or addons).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Category not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /product/addons:
    post:
      summary: Create a new addon
      tags: [Products]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAddonRequest'
      responses:
        '201':
          description: Addon created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Addon' # Returns full addon
        '400':
          description: Invalid input (e.g., name missing, parent category not found, name conflict).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Parent category not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict (e.g., addon with this name already exists in the parent category).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Get all addons
      tags: [Products]
      security:
        - cookieAuth: []
      parameters:
        - name: categoryId
          in: query
          required: false
          description: Filter addons by parent category ID.
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved addons.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Addon' # Returns full addons
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /product/addons/{addonId}:
    parameters:
      - name: addonId
        in: path
        required: true
        description: ID of the addon.
        schema:
          type: string
    get:
      summary: Get a specific addon by ID
      tags: [Products]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successfully retrieved the addon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Addon' # Returns full addon
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Addon not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update an addon
      tags: [Products]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAddonRequest'
      responses:
        '200':
          description: Addon updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Addon' # Returns updated full addon
        '400':
          description: Invalid input (e.g., parent category not found, name conflict).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Addon or target parent category not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict (e.g., addon with this name already exists in the target category).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete an addon
      tags: [Products]
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Addon deleted successfully.
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Addon not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...